+++ spec/project/default_config_spec.rb
@@ -24,7 +24,7 @@ describe 'config/default.yml' do
 
   it 'has a nicely formatted description for all cops' do
     cop_names.each do |name|
-      description = default_config[name]['Description']
+      description = default_config.fetch(name).fetch('Description')
       expect(description).not_to be_nil
       expect(description).not_to include("\n")
     end