+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add `AssignmentOnly` configuration option for `RSpec/InstanceVariable` cop. ([@backus][])
 * Add `BeEql` cop which looks for expectations that can use `be(...)` instead of `eql(...)`. ([@backus][])
 * Add autocorrect support for `BeEql` cop. ([@backus][])
+* Add `MessageExpectation` cop for enforcing consistent style of either `expect(...).to receive` or `allow(...).to receive`. ([@backus][])
 
 ## 1.6.0 (2016-08-03)